+++ src/natcap/invest/crop_production_percentile/__init__.py
@@ -1,0 +1,3 @@
+from .crop_production_percentile import execute
+from .crop_production_percentile import validate
+from .crop_production_percentile import MODEL_SPEC