+++ src/natcap/invest/seasonal_water_yield/seasonal_water_yield.py
@@ -2,22 +2,19 @@
 import fractions
 import logging
 import os
-import re
-import warnings
 
 import numpy
 import pygeoprocessing
 import pygeoprocessing.routing
 import scipy.special
-import taskgraph
 from osgeo import gdal
 from osgeo import ogr
 
-from .. import gettext
-from .. import spec
-from .. import utils
-from .. import validation
-from ..unit_registry import u
+from natcap.invest import gettext
+from natcap.invest import spec
+from natcap.invest import utils
+from natcap.invest import validation
+from natcap.invest.unit_registry import u
 from . import seasonal_water_yield_core
 
 LOGGER = logging.getLogger(__name__)