+++ src/natcap/invest/urban_flood_risk_mitigation/urban_flood_risk_mitigation.py
@@ -1,6 +1,5 @@
 """Urban Flood Risk Mitigation model."""
 import logging
-import os
 
 import numpy
 import pygeoprocessing
@@ -8,17 +7,13 @@ import rtree
 import scipy
 import shapely.prepared
 import shapely.wkb
-import taskgraph
 from osgeo import gdal
 from osgeo import ogr
-from osgeo import osr
 
-from . import gettext
-from . import spec
-from . import utils
-from . import validation
-from .unit_registry import u
-from .file_registry import FileRegistry
+from natcap.invest import gettext
+from natcap.invest import spec
+from natcap.invest import validation
+from natcap.invest.unit_registry import u
 
 LOGGER = logging.getLogger(__name__)