+++ src/natcap/invest/wind_energy/wind_energy.py
@@ -12,23 +12,19 @@ import shapely.ops
 import shapely.prepared
 import shapely.wkb
 import shapely.wkt
-import taskgraph
 from osgeo import gdal
 from osgeo import ogr
 from osgeo import osr
 from scipy import integrate
-from shapely import speedups
 from rtree import index
 
-from . import gettext
-from . import spec
-from . import utils
-from . import validation
-from .file_registry import FileRegistry
-from .unit_registry import u
+from natcap.invest import gettext
+from natcap.invest import spec
+from natcap.invest import utils
+from natcap.invest import validation
+from natcap.invest.unit_registry import u
 
 LOGGER = logging.getLogger(__name__)
-speedups.enable()
 
 WIND_DATA_FIELDS_FROM_INPUT = [
     spec.NumberOutput(